+++ Cargo.toml
@@ -1,25 +1,14 @@
-[package]
-name = "corespace-weigher"
-version = "0.1.0"
+[workspace.package]
+authors = ["RegionX <support@regionx.tech>"]
 edition = "2021"
+repository = "https://github.com/RegionX-Labs/CorespaceWeigher.git"
+license = "GPL-3.0-only"
 
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[[bin]]
-name = "consumption-tracker"
-path = "src/main.rs"
-
-[[bin]]
-name = "api"
-path = "src/api.rs"
-
-[dependencies]
-csv = "1.3.0"
-parity-scale-codec = "3.6.5"
-rocket = { version = "0.5.0", features=["json"] }
-rocket_cors = "0.6.0"
-serde = "1.0.193"
-serde_json = "1.0.108"
-subxt = "0.32.1"
-subxt-metadata = "0.32.1"
-tokio = { version = "1", features = ["full"] }
+[workspace]
+members = [
+    "bin/server",
+    "bin/tracker",
+    "routes",
+    "shared",
+    "types",
+]
