+++ bin/server/Cargo.toml
@@ -1,0 +1,12 @@
+[package]
+name = "server"
+version = "0.1.0"
+edition = "2021"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+rocket = { version = "0.5.0", features=["json"] }
+rocket_cors = "0.6.0"
+
+routes = { path = "../../routes" }
