+++ bin/tracker/Cargo.toml
@@ -1,0 +1,15 @@
+[package]
+name = "tracker"
+version = "0.1.0"
+edition = "2021"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+csv = "1.3.0"
+subxt = "0.32.1"
+subxt-metadata = "0.32.1"
+tokio = { version = "1", features = ["full"] }
+
+types = { path = "../../types" }
+shared = { path = "../../shared" }
