+++ routes/src/consumption.rs
@@ -1,0 +1,1 @@
+
