+++ types/Cargo.toml
@@ -1,0 +1,14 @@
+[package]
+name = "types"
+version = "0.1.0"
+authors.workspace = true
+edition.workspace = true
+repository.workspace = true
+license.workspace = true
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+serde = "1.0.193"
+rocket = { version = "0.5.0", features=["json"] }
+
